# openapi.yaml
openapi: 3.0.3
info:
  title: Hava Durumu API
  version: "1.0.0"
  description: >
    MGM (Meteoroloji Genel Müdürlüğü) verisini kullanan ve resmi olmayan bir REST
    API. Redis destekli önbellek, stale-while-revalidate ve MGM kesintilerine
    karşı circuit breaker içerir bkz. README.md
  license:
    name: MIT

servers:
  - url: http://localhost:5000
    description: Yerel geliştirme

tags:
  - name: iller
    description: İl/ilçe ve istasyon bilgisi
  - name: hava-durumu
    description: Güncel durum, tahmin, birleşik özet
  - name: uyarilar
    description: Meteorolojik uyarılar (deneysel, ham geçiş)
  - name: durum
    description: Servis sağlığı

paths:
  /iller:
    get:
      tags: [iller]
      summary: Türkiye'nin 81 ilini listeler
      description: >
        Sabit veridir; MGM'ye istek atmaz, önbellek/circuit breaker'dan
        etkilenmez. Plaka kodu sırasıyla döner.
      responses:
        "200":
          description: İl listesi
          content:
            application/json:
              schema:
                type: object
                properties:
                  basarili: { type: boolean, example: true }
                  veri:
                    type: array
                    items:
                      $ref: "#/components/schemas/Il"

  /ara:
    get:
      tags: [hava-durumu]
      summary: Serbest metin sorgusuyla hava durumu (yazım hatası/karmaşık girdi)
      description: >
        `"kadıköy"`, `"kadikoy"`, `"kadikoy/istanbul"`, `"maslak itü"` gibi
        eksik veya hatalı Türkçe karakterli ya da il+ilçe'yi tek metinde birleşik
        veren girdileri çözer. Katmanlı çalışır: önce 81 il listesiyle
        tam eşleşmeye bakar, sonra `/`, `,` veya boşlukla
        ayrılmış parçalardan il+ilçe çıkarıp MGM'ye doğrudan sorar, hâlâ
        çözemezse key gerektirmeyen Open-Meteo geocoding servisine düşer. 
        Yanıttaki `yontem` alanı hangi katmanın çözdüğünü gösterir.
      parameters:
        - name: q
          in: query
          required: true
          description: Serbest metin yer sorgusu
          schema:
            type: string
            example: "maslak itü"
      responses:
        "200":
          description: >
            Çözüldü ya da belirsiz
          content:
            application/json:
              schema:
                type: object
                properties:
                  basarili: { type: boolean, example: true }
                  veri:
                    oneOf:
                      - $ref: "#/components/schemas/AramaCozuldu"
                      - $ref: "#/components/schemas/AramaBelirsiz"
        "400":
          description: "'q' parametresi eksik ya da boş"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HataYaniti"
        "404":
          description: Sorgu hiçbir yere çözümlenemedi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HataYaniti"

  /uyarilar:
    get:
      tags: [uyarilar]
      summary: Meteorolojik uyarılar (deneysel — MGM şeması doğrulanmadı)
      description: >
        MGM'nin meteorolojik uyarı (MeteoUYARI) verisini döndürür. **Önemli
        sınır:** bu uç nokta diğerlerinden farklıdır — MGM'nin ham JSON
        şemasını, bu API'nin geliştirildiği sırada Türkiye genelinde aktif
        bir uyarı bulunmadığı için doğrulayamadık. Bu yüzden veri
        dönüştürülmeden, MGM'den geldiği gibi `ham` alanı altında döner;
        alan adlarını tahmin ederek Türkçeleştirme/yeniden şekillendirme
        yapılmaz. MGM'nin resmi MeteoUYARI sistemi şiddet için
        Yeşil/Sarı/Turuncu/Kırmızı, hadise tipi için Soğuk/Sıcak/Sis/Zirai
        Don/Buzlanma ve Don/Toz Taşınımı/Kar Erimesi/Çığ/Kar/Gökgürültülü
        Sağanak Yağış/Rüzgar/Yağmur kategorilerini kullanıyor (bkz.
        mgm.gov.tr/meteouyari) — ama bu kategorilerin MGM'nin JSON
        yanıtındaki gerçek alan adları doğrulanmadan varsayılmadı.
      parameters:
        - name: il
          in: query
          required: false
          description: >
            MGM'ye doğrudan parametre olarak iletilir. Filtrenin MGM
            tarafında gerçekten desteklenip desteklenmediği doğrulanamadı
            (aktif uyarı olmadığından test edilemedi) — zararsız bir
            passthrough'tur.
          schema:
            type: string
            example: istanbul
      responses:
        "200":
          description: MGM'nin ham uyarı verisi (dönüştürülmemiş)
          content:
            application/json:
              schema:
                type: object
                properties:
                  basarili: { type: boolean, example: true }
                  veri:
                    type: object
                    properties:
                      ham:
                        type: array
                        description: MGM'den geldiği gibi, dönüştürülmemiş dizi. Aktif uyarı yoksa boş.
                        items: { type: object }
                      not:
                        type: string
                        description: Şemanın neden doğrulanamadığını açıklayan not.
        "502":
          description: MGM'ye ulaşılamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HataYaniti"

  /konum:
    get:
      tags: [hava-durumu]
      summary: GPS koordinatından hava durumu
      description: >
        Enlem ve boylamı OpenStreetMap'in ücretsiz Nominatim servisiyle
        bir il/ilçe adayına çevirir. MGM'de karşılığı yoksa
        doğrudan verilen koordinatla Open-Meteo'dan güncel durum döner
      parameters:
        - name: lat
          in: query
          required: true
          description: Enlem, -90 ile 90 arası
          schema:
            type: number
            example: 40.99
        - name: lon
          in: query
          required: true
          description: Boylam, -180 ile 180 arası
          schema:
            type: number
            example: 29.02
      responses:
        "200":
          description: Çözüldü
          content:
            application/json:
              schema:
                type: object
                properties:
                  basarili: { type: boolean, example: true }
                  veri:
                    $ref: "#/components/schemas/KonumCozuldu"
        "400":
          description: "'lat'/'lon' eksik, sayı değil ya da aralık dışı"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HataYaniti"
        "502":
          description: Hem MGM hem Open-Meteo'ya ulaşılamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HataYaniti"

  /istasyonlar/{il}:
    get:
      tags: [iller]
      summary: Bir ildeki tüm MGM istasyonlarını (ilçelerini) listeler
      parameters:
        - $ref: "#/components/parameters/IlPath"
      responses:
        "200":
          description: İstasyon listesi
          content:
            application/json:
              schema:
                type: object
                properties:
                  basarili: { type: boolean, example: true }
                  veri:
                    type: array
                    items:
                      $ref: "#/components/schemas/Istasyon"
        "404":
          $ref: "#/components/responses/BulunamadiHatasi"

  /guncel/{il}:
    get:
      tags: [hava-durumu]
      summary: Anlık (güncel) hava durumu
      parameters:
        - $ref: "#/components/parameters/IlPath"
        - $ref: "#/components/parameters/IlceQuery"
      responses:
        "200":
          description: Güncel durum
          content:
            application/json:
              schema:
                type: object
                properties:
                  basarili: { type: boolean, example: true }
                  veri:
                    $ref: "#/components/schemas/GuncelDurum"
        "404":
          $ref: "#/components/responses/BulunamadiHatasi"

  /tahmin/{il}:
    get:
      tags: [hava-durumu]
      summary: 5 günlük hava tahmini
      parameters:
        - $ref: "#/components/parameters/IlPath"
        - $ref: "#/components/parameters/IlceQuery"
      responses:
        "200":
          description: Günlük tahmin listesi (5 gün)
          content:
            application/json:
              schema:
                type: object
                properties:
                  basarili: { type: boolean, example: true }
                  veri:
                    type: array
                    items:
                      $ref: "#/components/schemas/GunlukTahmin"
        "404":
          $ref: "#/components/responses/BulunamadiHatasi"

  /saatlik/{il}:
    get:
      tags: [hava-durumu]
      summary: Saatlik hava tahmini (MGM'de mevcutsa)
      description: >
        MGM'nin ham saatlik tahmin kaydını döner; alanlar istasyona göre
        değişebileceğinden şema serbest formatlıdır.
      parameters:
        - $ref: "#/components/parameters/IlPath"
        - $ref: "#/components/parameters/IlceQuery"
      responses:
        "200":
          description: Saatlik tahmin listesi
          content:
            application/json:
              schema:
                type: object
                properties:
                  basarili: { type: boolean, example: true }
                  veri:
                    type: array
                    items:
                      type: object
        "404":
          $ref: "#/components/responses/BulunamadiHatasi"

  /hava-durumu/{il}:
    get:
      tags: [hava-durumu]
      summary: Güncel durum + 5 günlük tahmin + gün doğumu/batımı (birleşik)
      description: >
        `/guncel`, `/tahmin` ve gün doğumu/batımı verisini tek istekte
        toplayan üst düzey uç nokta. Gün doğumu/batımı verisi ayrı bir dış
        servisten geldiği için başarısız olursa (sessizce) atlanır; diğer
        alanlar yine döner.
      parameters:
        - $ref: "#/components/parameters/IlPath"
        - $ref: "#/components/parameters/IlceQuery"
      responses:
        "200":
          description: Birleşik hava durumu özeti
          content:
            application/json:
              schema:
                type: object
                properties:
                  basarili: { type: boolean, example: true }
                  veri:
                    $ref: "#/components/schemas/HavaDurumuOzeti"
        "404":
          $ref: "#/components/responses/BulunamadiHatasi"

  /health:
    get:
      tags: [durum]
      summary: Servis sağlığı
      description: >
        `deep` parametresiz: sadece process'in ayakta olduğunu
        doğrular, MGM/Redis'e istek atmaz ve rate limit'e uğramaz.
        `?deep=1` gerçekten MGM'ye bir istek atar ve Redis bağlantısını
        kontrol eder buna göre MGM veya Redis çökükse 503 döner.
      parameters:
        - name: deep
          in: query
          required: false
          description: "1/true/yes/on değerlerinden biriyse derin kontrol yapılır."
          schema:
            type: string
            example: "1"
      responses:
        "200":
          description: Servis sağlıklı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaglikDurumu"
        "503":
          description: MGM veya Redis'e ulaşılamıyor (yalnızca deep=1 iken)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaglikDurumu"

components:
  parameters:
    IlPath:
      name: il
      in: path
      required: true
      description: "İl adı. Türkçe karakter sorunu yaşanırsa aksansız yazın (Istanbul, Bakirkoy)."
      schema:
        type: string
        example: Istanbul
    IlceQuery:
      name: ilce
      in: query
      required: false
      description: >
        Belirtilmezse MGM'nin o il için döndürdüğü varsayılan istasyon
        kullanılır. Not: `/istasyonlar/<il>` MGM'nin kısıtı yüzünden o
        ilin tüm ilçelerini listeleyemez (MGM bunu desteklemiyor) — geçerli
        bir ilçe adını zaten biliyorsanız doğrudan burada belirtin, MGM'ye
        ayrıca sorgulanır.
      schema:
        type: string
        example: Bakirkoy

  responses:
    BulunamadiHatasi:
      description: İl/ilçe/istasyon bulunamadı ya da MGM'den veri alınamadı
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HataYaniti"

  schemas:
    Il:
      type: object
      properties:
        plakaKodu: { type: integer, example: 34 }
        il: { type: string, example: İstanbul }

    Istasyon:
      type: object
      description: MGM'nin ham istasyon/merkez kaydı; alan adları istasyona göre değişebilir.
      properties:
        istasyonId: { type: integer, example: 17062 }
        merkezId: { type: integer, example: 17062 }
        il: { type: string, example: İstanbul }
        ilce: { type: string, example: Bakırköy }
        enlem: { type: number, example: 40.98 }
        boylam: { type: number, example: 28.87 }
      additionalProperties: true

    GuncelDurum:
      type: object
      description: >
        `kaynak: "mgm"` normal yanıttır. MGM'ye ulaşılamadığında (circuit
        breaker açık dahil) `kaynak: "open-meteo"` ile ücretsiz, key
        gerektirmeyen bir yedek servisten gelen veri döner — `durumKodu` bu
        durumda MGM'nin değil, Open-Meteo'nun WMO kod uzayındandır, ikisi
        karşılaştırılamaz.
      properties:
        istasyonId: { type: integer, example: 17062 }
        sicaklik: { type: number, nullable: true, example: 21.4 }
        nem: { type: number, nullable: true, example: 62 }
        ruzgarHizi: { type: number, nullable: true, example: 11.2 }
        ruzgarYonu: { type: number, nullable: true, example: 180 }
        basinc: { type: number, nullable: true, example: 1013.2 }
        denizSeviyesiBasinc: { type: number, nullable: true, example: 1015.1 }
        durumKodu:
          nullable: true
          example: 1
          description: "kaynak=mgm iken MGM kodu (string), kaynak=open-meteo iken WMO kodu (integer)"
        durum: { type: string, nullable: true, example: Açık }
        olcumZamani: { type: string, nullable: true, example: "2026-08-15 09:00:00" }
        kaynak:
          type: string
          enum: [mgm, open-meteo]
          example: mgm
          description: Verinin hangi servisten geldiği.

    GunlukTahmin:
      type: object
      properties:
        tarih: { type: string, format: date, example: "2026-08-15" }
        enDusuk: { type: number, nullable: true, example: 18.0 }
        enYuksek: { type: number, nullable: true, example: 29.5 }
        enDusukNem: { type: number, nullable: true, example: 45 }
        enYuksekNem: { type: number, nullable: true, example: 70 }
        ruzgarHizi: { type: number, nullable: true, example: 9.5 }
        durumKodu: { type: integer, nullable: true, example: 1 }
        durum: { type: string, nullable: true, example: Açık }

    HavaDurumuOzeti:
      type: object
      properties:
        il: { type: string, example: İstanbul }
        ilce: { type: string, nullable: true, example: Bakırköy }
        istasyonId: { type: integer, example: 17062 }
        enlem: { type: number, nullable: true, example: 40.98 }
        boylam: { type: number, nullable: true, example: 28.87 }
        guncel: { $ref: "#/components/schemas/GuncelDurum" }
        tahmin:
          type: array
          items: { $ref: "#/components/schemas/GunlukTahmin" }
        gunDogumu: { type: string, nullable: true, example: "06:12" }
        gunBatimi: { type: string, nullable: true, example: "19:48" }

    AramaCozuldu:
      allOf:
        - $ref: "#/components/schemas/HavaDurumuOzeti"
        - type: object
          properties:
            durum: { type: string, enum: [cozuldu], example: cozuldu }
            sorgu: { type: string, example: "maslak itü" }
            yontem:
              type: string
              enum: [il-eslesme, il-ilce-parcalama, geocoding-mgm, geocoding-dogrudan]
              description: >
                il-eslesme/il-ilce-parcalama: MGM üzerinden tam hava_durumu
                yanıtı. geocoding-mgm: Open-Meteo geocoding ile bulunan
                il/ilçe MGM'de de doğrulandı. geocoding-dogrudan: MGM'de
                karşılığı yok (ör. resmi ilçe olmayan bir mahalle),
                `guncel` doğrudan Open-Meteo'dan; `tahmin` boş liste.
              example: geocoding-dogrudan

    KonumCozuldu:
      allOf:
        - $ref: "#/components/schemas/HavaDurumuOzeti"
        - type: object
          properties:
            durum: { type: string, enum: [cozuldu], example: cozuldu }
            yontem:
              type: string
              enum:
                [
                  nominatim-mgm,
                  nominatim-mgm-il-varsayilan,
                  nominatim-open-meteo,
                  open-meteo-dogrudan,
                ]
              description: >
                nominatim-mgm/nominatim-mgm-il-varsayilan: Nominatim ile
                bulunan il(+ilçe) MGM'de de doğrulandı. 
                nominatim-open-meteo: Nominatim bir adres buldu ama MGM'de karşılığı yoktu.
                open-meteo-dogrudan: Nominatim hiç adres bulamadı ya da servise ulaşılamadı.
              example: nominatim-mgm

    AramaBelirsiz:
      type: object
      properties:
        durum: { type: string, enum: [belirsiz], example: belirsiz }
        sorgu: { type: string, example: "merkez mahallesi" }
        secenekler:
          type: array
          description: Farklı illerde birden fazla makul aday bulundu; tahmin yürütülmedi.
          items:
            type: object
            properties:
              yer: { type: string, example: Merkez }
              il: { type: string, example: Konya }
              ulke: { type: string, example: Türkiye }
              enlem: { type: number, example: 37.87 }
              boylam: { type: number, example: 32.49 }

    SaglikDurumu:
      type: object
      properties:
        basarili: { type: boolean }
        durum: { type: string, enum: [ok, degraded], example: ok }
        servis: { type: string, example: hava-durumu }
        mgm: { type: string, enum: [ok, hata, skip], example: skip }
        redis: { type: string, enum: [ok, hata, skip], example: skip }
        circuit_breaker: { type: string, enum: [kapali, acik, yari-acik], example: kapali }
        hata: { type: string, nullable: true }

    HataYaniti:
      type: object
      properties:
        basarili: { type: boolean, example: false }
        hata: { type: string, example: "'Xyz' için istasyon bulunamadı." }